Add lib folder to scss
The lib/_lib.scss file now acts as the general importer for
any external css libraries that might be needed. For instance,
if you are using jQuery-ui, you could rename that css file to
scss and include it in the libs folder, then add an import in
_lib.scss. It will be bundled and minified with the rest when
the compile is run.